requirements.txt
requests==2.21.0
Pillow==6.0.0
Click==7.0
future==0.17.1
huepy==0.9.8.1
instabot==0.4.5
certifi==2019.3.9
idna==2.8
itsdangerous==1.1.0
PySocks==1.6.8
requests-toolbelt==0.9.1
schedule==0.6.0
tqdm==4.31.1
python-dotenv==0.10.1